+++ docs/CHANGELOG.md
@@ -56,6 +56,7 @@ Version 1.1.0 (Tamarack): Not yet released
 - Add `dbg_mapver` command to check RFL versions
 - Add `sv_connectedclients` command to help servers understand which clients can load which maps
 - Add notification to players when the server is about to switch to a map with a version they can't load
+- Add `mapm` command as an alias for `levelm`
 
 
 Version 1.0.0 (Maple): Released Jan-25-2025